bloxpkgs: reworking custom package handling
By adding each package to a nixpkgs overlay they become available to
tools like nix-env and nix-shell as overlays are propagated down to
users via the previous commit.

I am still reluctant to merge the customized pacakges to the nixpkgs
tree so they are all collected under the `bloxpkgs` subtree.